g4-common: wrild: fixes & introducing watchdog
global:

- added developer debug option (everything but no ril restarts)
- fix: SIM-less mode in rild and watchdog
- complete re-work of dontaudit (macro implemented)

rild related fixes/enhancements:

- faster detection of ril issues
- full support for encrypted devices
- more reliable boot process detection
- fix: issue when READY but no operator
- better logging in some places
- RILRESTART moved to a function
- fix: do not count when on PIN screen
- fix: counter wasn't reset in all cases

watchdog related:

before: wrild went to background forever.
now: wrild will go to background but loading a watchdog
which will run forever.

This introduces a (hopefully) intelligent watchdog for rild.
Main reason was high CPU load in very rare cases.
The watchdog is fully integrated into wrild and so makes
use of e.g. the new RILRESTART function when triggering
a restart is needed.

watchdog main features:

- actions of this watchdog begin with "woof" (adb logcat -s WRILD -e woof)
- configurable: DEBUGLOG on/off (default: /sdcard/Download/wdlog)
- configurable: CPU threshold
- configurable: max duration of breaking the CPU threshold
- configurable: watchdog frequency
- configurable: 3rd party apps which provide in-app-calls
(when in foreground: pausing, when in background and active call: pausing)

watchdog process:

The watchdog checks (intervall: WDFREQ) if the CPU and max retries thresholds
are exceeded for the defined duration.
If thats the case it will check for the following and do NOTHING when any of those
are met:

- active call (dialer) in foreground or background
- ringing (dialer)
- 3rd party call app (CALLAPPS) in foreground
- 3rd party call app (CALLAPPS) in background AND in-call

If NONE of these apply 3 logfiles get created (in DOGLOGS):

- YYYY-MM-DD_<rild-PID>_logcat.txt
- YYYY-MM-DD_<rild-PID>_dmesg.txt
- YYYY-MM-DD_<rild-PID>_ps.txt

Afterwards rild gets kicked out of the field and restarted properly.
That means we also check if the cell service coming back as it should.
If not: we restart RILD again like we did on the first boot (when necessary).

grab debug logs:

newer adb versions support pulling a complete dir so if you left DOGLOGS
at default this will do:

adb pull /sdcard/Download/wdlog